+++ pylint/lint/parallel.py
@@ -3,7 +3,7 @@
 
 import collections
 import functools
-from typing import Any, DefaultDict, Iterator, List, Tuple
+from typing import Any, DefaultDict, Iterable, List, Tuple
 
 from pylint import reporters
 from pylint.lint.utils import _patch_sys_path
@@ -67,12 +67,10 @@ def _worker_initialize(linter, arguments=None):
 def _worker_check_single_file(
     file_item: FileItem,
 ) -> Tuple[int, Any, str, Any, List[Tuple[Any, ...]], Any, Any, DefaultDict[Any, List]]:
-    name, filepath, modname = file_item
-
     if not _worker_linter:
         raise Exception("Worker linter not yet initialised")
     _worker_linter.open()
-    _worker_linter.check_single_file(name, filepath, modname)
+    _worker_linter.check_single_file(file_item)
     mapreduce_data = collections.defaultdict(list)
     for checker in _worker_linter.get_checkers():
         try:
@@ -85,7 +83,7 @@ def _worker_check_single_file(
     return (
         id(multiprocessing.current_process()),
         _worker_linter.current_name,
-        filepath,
+        file_item[1],
         _worker_linter.file_state.base_name,
         msgs,
         _worker_linter.stats,
@@ -115,7 +113,7 @@ def _merge_mapreduce_data(linter, all_mapreduce_data):
             checker.reduce_map_data(linter, collated_map_reduce_data[checker.name])
 
 
-def check_parallel(linter, jobs, files: Iterator[FileItem], arguments=None):
+def check_parallel(linter, jobs, files: Iterable[FileItem], arguments=None):
     """Use the given linter to lint the files with given amount of workers (jobs)
     This splits the work filestream-by-filestream. If you need to do work across
     multiple files, as in the similarity-checker, then inherit from MapReduceMixin and
